Move F-case specific subroutines into separate files. Update dat2nc.c to store requests without defining variables with additional space.
