+++ Makefile
@@ -1,8 +1,10 @@
 
 MPICC		= mpicc
 CFLAGS          = -O2
+CFLAGS          = -O0 -g
 
 PnetCDF_DIR	= $(HOME)/PnetCDF/1.11.0
+PnetCDF_DIR	= $(HOME)/PnetCDF/GitHub/dest
 
 INCLUDES	= -I$(PnetCDF_DIR)/include -I.
 LDFLAGS		= -L$(PnetCDF_DIR)/lib
@@ -16,10 +18,15 @@ all: e3sm_io
 dat2nc: dat2nc.o
 	$(MPICC) $(CFLAGS) -o $@ $^ $(LDFLAGS) $(LIBS)
 
-e3sm_io.o: e3sm_io.h
-def_header.o: e3sm_io.h
+e3sm_io.o: e3sm_io.c e3sm_io.h
+read_decomp.o: read_decomp.c e3sm_io.h
 
-e3sm_io: e3sm_io.o def_header.o
+header_io_F_case.o: header_io_F_case.c e3sm_io.h
+var_io_F_case.o: var_io_F_case.c e3sm_io.h
+
+OBJS = read_decomp.o header_io_F_case.o var_io_F_case.o
+
+e3sm_io: e3sm_io.o $(OBJS)
 	$(MPICC) $(CFLAGS) -o $@ $^ $(LDFLAGS) $(LIBS)
 
 # romio_patch.c contains fix in https://github.com/pmodels/mpich/pull/3089
@@ -27,7 +34,7 @@ e3sm_io: e3sm_io.o def_header.o
 # ROMIO_PATCH	= -Wl,--wrap=ADIOI_Type_create_hindexed_x -l:libmpi.a
 ROMIO_PATCH	= -Wl,--wrap=ADIOI_Type_create_hindexed_x
 
-e3sm_io.romio_patch: e3sm_io.o def_header.o romio_patch.o
+e3sm_io.romio_patch: e3sm_io.o header_io_case_F.o romio_patch.o
 	$(MPICC) $(CFLAGS) -o $@ $^ $(LDFLAGS) $(LIBS) $(ROMIO_PATCH)
 
 romio_patch.o: romio_patch.c
@@ -35,8 +42,8 @@ romio_patch.o: romio_patch.c
 
 clean:
 	rm -f core.* *.o dat2nc e3sm_io e3sm_io.romio_patch
-	rm -f testfile_h0_varn.nc testfile_h1_varn.nc
-	rm -f testfile_h0_vard.nc testfile_h1_vard.nc
+	rm -f f_case_h0_varn.nc f_case_h1_varn.nc
+	rm -f f_case_h0_vard.nc f_case_h1_vard.nc
 
 .PHONY: clean
