+++ dat2nc.c
@@ -18,23 +18,35 @@
 #include <mpi.h>
 #include <pnetcdf.h>
 
-#define LINE_SIZE 2097152
+#define LINE_SIZE 4692802
 
 #define ERR {if(err!=NC_NOERR){printf("Error at line %d in %s: %s\n", __LINE__,__FILE__, ncmpi_strerrno(err));nerrs++;}}
 
 static int verbose, line_sz;
 
+/*----< intcompare() >------------------------------------------------------*/
+/* This subroutine is used in qsort() */
+static int intcompare(const void *p1, const void *p2)
+{
+    int i = *((int *)p1);
+    int j = *((int *)p2);
+    if (i > j) return (1);
+    if (i < j) return (-1);
+    return (0);
+}
+
 /*----< add_decomp() >------------------------------------------------------------*/
-int add_decomp(int   ncid,
-               char *infname,
-               char *label)
+static int
+add_decomp(int         ncid,
+           const char *infname,
+           int         label)
 {
     char  *buf, name[128], *map, *str;
     FILE *fd;
     int   rank, nprocs, ndims;
-    int   i, j, dimids[2], varid[2], *nreqs, err, nerrs=0, *off;
-    int max_nreqs, min_nreqs;
-    MPI_Offset k, gsize, *dims, start[2], count[2], dim_len;
+    int   i, j, dimid, varid[3], *nreqs, err, nerrs=0, *off, *len;
+    int   total_nreqs, max_nreqs, min_nreqs;
+    MPI_Offset k, gsize, *dims, start, count;
 
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
 
@@ -45,11 +57,16 @@ int add_decomp(int   ncid,
 	exit(1);
     }
 
+    /* buffer stores one line read from input file */
     buf = (char*) malloc(line_sz);
 
-    /* process header lines:
-     * version 2001 npes 43200 ndims 2 
-     * 777602 72 
+    /* header lines: (first 2 lines of the decomposition file), for example
+     *     version 2001 npes 43200 ndims 2
+     *     777602 72
+     * version is 2001
+     * number of MPI processes used to generate this decomposition is 43200
+     * number of dimensions of decomposed variable is 2
+     * 1st dimension is of size 777602 and 2nd is 72 (in Fortran order)
      */
     fgets(buf, LINE_SIZE, fd);
     if (verbose) printf("header:\n\t%s",buf);
@@ -67,7 +84,7 @@ int add_decomp(int   ncid,
     nprocs = atoi(strtok(NULL, " "));
     strtok(NULL, " "); /* token "ndims" */
     ndims = atoi(strtok(NULL, " "));
-    dims = (MPI_Offset*) malloc(3 * ndims * sizeof(MPI_Offset));
+    dims = (MPI_Offset*) malloc(ndims * sizeof(MPI_Offset));
 
     /* get dimension sizes
      * Note the dimensions are in Fortran order in the decomposition file
@@ -75,33 +92,76 @@ int add_decomp(int   ncid,
      */
     fgets(buf, LINE_SIZE, fd);
     if (verbose) printf("\t%s",buf);
-    dims[ndims-1] = atoll(strtok(buf, " "));
-    for (i=ndims-2; i>=0; i--)
+    dims[0] = atoll(strtok(buf, " "));
+    for (i=1; i<ndims; i++)
         dims[i] = atoll(strtok(NULL, " "));
+    /* Note dims[] is in Fortran order */
+    if (verbose) {
+        if (ndims == 1) printf("lable D%d: dims = %lld\n",label, dims[0]);
+        if (ndims == 2) printf("lable D%d: dims = %lld %lld\n",label, dims[0], dims[1]);
+    }
 
+    /* gsize is total number of elements in the global array */
     gsize=dims[0];
     for (i=1; i<ndims; i++) gsize *= dims[i];
-    
+
+    /* map is used to check whether the entire array is covered by requests
+     * of all processes.
+     */
     map = (char*) calloc(gsize, 1);
+
+    /* nreqs[i] is the number of elements accessed by process i */
     nreqs = (int*) malloc(nprocs * sizeof(int));
 
-    /* calculate max_nreqs, min_nreqs and nreqs[] */
+    /* calculate max_nreqs, min_nreqs and nreqs[]
+     * decomposition data format:
+     *     (process.rank.ID) (number.of.requests)
+     *     a list of element offsets accessed by this process (one element each)
+     * Offsets are indices of array flattened into 1D and start with 1, i.e.
+     * Fortran index based. Note the offsets are not sorted in an increasing
+     * order and may contain 0s which should be ignored.
+     */
+    total_nreqs = 0;
     for (rank=0; rank<nprocs; rank++) {
-        fgets(buf, LINE_SIZE, fd);
-        assert(rank == atoi(strtok(buf, " ")));
-        nreqs[rank] = atoi(strtok(NULL, " "));
+        int ncontig=0, decomp_rank;
+        char *ret;
+
+        /* reads the first line of rank ID and no. requests */
+        while (NULL != (ret = fgets(buf, LINE_SIZE, fd))) {
+            if (buf[0] != '\n') break; /* non-empty line */
+        }
+        if (ret == NULL || strncmp(buf, "Obtained", 8) == 0) {
+            /* there is no request for remaining ranks */
+            for (decomp_rank=rank; decomp_rank<nprocs; decomp_rank++)
+                nreqs[rank] = 0;
+            min_nreqs = 0;
+            break; /* loop of rank */
+        }
+
+        decomp_rank = atoi(strtok(buf, " ")); /* rank ID */
+        while (rank < decomp_rank) { /* this rank has no request */
+            nreqs[rank++] = 0;
+            min_nreqs = 0;
+        }
+
+        nreqs[rank] = atoi(strtok(NULL, " "));  /* number of requests */
+        if (nreqs[rank] == 0) { /* this rank has zero request */
+            min_nreqs = 0;
+            continue; /* loop of rank */
+        }
+
         off = (int*) malloc(nreqs[rank] * sizeof(int));
-        fgets(buf, LINE_SIZE+1, fd);
+        fgets(buf, LINE_SIZE+1, fd); /* 2nd line: list of offsets */
         if (strlen(buf) >= LINE_SIZE) {
             printf("Error: line size is larger than default %d\n",LINE_SIZE);
             printf("       use command-line option -l to use a larger size\n");
             goto fn_exit;
         }
 
-        /* read offset list */
+        /* construct the offset list */
         off[0] = atoi(strtok(buf, " "));
         j=1;
-        while (off[0] == 0) {
+        while (off[0] == 0) { /* skip leading 0 values, if there is any */
             off[0] = atoi(strtok(NULL, " "));
             j++;
         }
@@ -113,99 +173,135 @@ int add_decomp(int   ncid,
             off[k]--;  /* offset is 1 based */
             k++;
         }
-        if (rank == 0) {
-            max_nreqs = min_nreqs = k;
+
+        /* sort off[] into an increasing order */
+        qsort((void*)off, k, sizeof(int), intcompare);
+
+        ncontig = 1;
+        for (j=1; j<k; j++) {
+            /* break contiguity at dimension boundaries or noncontiguous */
+            if (off[j] % dims[0] == 0 || off[j] > off[j-1] + 1)
+                ncontig++;
         }
-        else {
-            max_nreqs = (k > max_nreqs) ? k : max_nreqs;
-            min_nreqs = (k < min_nreqs) ? k : min_nreqs;
-        }
-        nreqs[rank]=k;
 
         for (j=0; j<k; j++) map[off[j]]=1;
         free(off);
+
+        total_nreqs += ncontig;
+
+        /* find max and min nreqs amount all processes */
+        if (rank == 0) {
+            max_nreqs = min_nreqs = ncontig;
+        }
+        else {
+            max_nreqs = (ncontig > max_nreqs) ? ncontig : max_nreqs;
+            min_nreqs = (ncontig < min_nreqs) ? ncontig : min_nreqs;
+        }
+        nreqs[rank] = ncontig;
     }
-    if (verbose) printf("max_nreqs=%d min_nreqs=%d\n",max_nreqs, min_nreqs);
+
+    if (verbose)
+        printf("max_nreqs=%d min_nreqs=%d\n",max_nreqs, min_nreqs);
+
+    /* check if the entire array is covered */
     for (k=0,j=0; j<gsize; j++) k+=map[j];
     if (k != gsize)
         printf("Warning: global array size = %lld but only %lld written\n",gsize,k);
     free(map);
 
-    /* check if dimension num_procs has been defined */
-    err = ncmpi_inq_dimid(ncid, "num_procs", &dimids[0]);
+    /* check if dimension decomp_nprocs has been defined in the netCDF file */
+    err = ncmpi_inq_dimid(ncid, "decomp_nprocs", &dimid);
     if (err == NC_EBADDIM) {  /* not defined */
-        err = ncmpi_def_dim(ncid, "num_procs", nprocs, &dimids[0]); ERR
+        err = ncmpi_def_dim(ncid, "decomp_nprocs", nprocs, &dimid); ERR
     }
     else {
-        MPI_Offset num_procs;
-        err = ncmpi_inq_dimlen(ncid, dimids[0], &num_procs); ERR
-        if (num_procs != nprocs) {
-            printf("Error: num_procs mismatches among input files\n");
+        /* if decomp_nprocs already exist, check if value matches */
+        MPI_Offset decomp_nprocs;
+        err = ncmpi_inq_dimlen(ncid, dimid, &decomp_nprocs); ERR
+        if (decomp_nprocs != nprocs) {
+            printf("Error: decomp_nprocs=%lld mismatches among input files %d\n", decomp_nprocs, nprocs);
             MPI_Finalize();
             exit(1);
         }
     }
 
-    sprintf(name, "%s.max_nreqs", label);
-    err = ncmpi_def_dim(ncid, name, max_nreqs, &dimids[1]); ERR
-    sprintf(name, "%s.nreqs", label);
-    err = ncmpi_def_var(ncid, name, NC_INT, 1, dimids, &varid[0]); ERR
-    str = "Number of noncontiguous subarray requests by each MPI process";
+    /* define variable nreqs for this decomposition */
+    sprintf(name, "D%d.nreqs", label);
+    err = ncmpi_def_var(ncid, name, NC_INT, 1, &dimid, &varid[0]); ERR
+
+    /* define attribute description for this variable */
+    str = "Number of noncontiguous requests per process";
     err = ncmpi_put_att_text(ncid, varid[0], "description", strlen(str), str); ERR
 
-    sprintf(name, "%s.offsets", label);
-    err = ncmpi_def_var(ncid, name, NC_INT, 2, dimids, &varid[1]); ERR
-    str = "Flattened starting indices of noncontiguous requests. Each row corresponds to requests by an MPI process.";
+    /* define dimension total_nreqs for this decomposition */
+    sprintf(name, "D%d.total_nreqs", label);
+    err = ncmpi_def_dim(ncid, name, total_nreqs, &dimid); ERR
+
+    /* define variable offsets (store starting element indices) */
+    sprintf(name, "D%d.offsets", label);
+    err = ncmpi_def_var(ncid, name, NC_INT, 1, &dimid, &varid[1]); ERR
+    str = "Flattened starting indices of noncontiguous requests";
     err = ncmpi_put_att_text(ncid, varid[1], "description", strlen(str), str); ERR
 
-    if (ndims > 1) {
-        err = ncmpi_get_att_longlong(ncid, NC_GLOBAL, "dim_len_Y", &dim_len);
-        if (err == NC_ENOTATT) {
-            err = ncmpi_put_att_longlong(ncid, NC_GLOBAL, "dim_len_Y", NC_INT, 1, &dims[0]);
-            ERR
-        }
-        else {
-            if (dim_len != dims[0]) {
-                printf("Error: dim_len_Y mismatches among input files\n");
-                MPI_Finalize();
-                exit(1);
-            }
-        }
-    }
-    err = ncmpi_get_att_longlong(ncid, NC_GLOBAL, "dim_len_X", &dim_len);
-    if (err == NC_ENOTATT) {
-        dim_len = (ndims == 1) ? dims[0] : dims[1];
-        err = ncmpi_put_att_longlong(ncid, NC_GLOBAL, "dim_len_X", NC_INT, 1, &dim_len);
+    /* define variable lengths (store number of elements) */
+    sprintf(name, "D%d.lengths", label);
+    err = ncmpi_def_var(ncid, name, NC_INT, 1, &dimid, &varid[2]); ERR
+    str = "Lengths of noncontiguous requests";
+    err = ncmpi_put_att_text(ncid, varid[2], "description", strlen(str), str); ERR
+
+    /* add attribute to describe dimensionality */
+    sprintf(name, "D%d.ndims", label);
+    err = ncmpi_put_att_int(ncid, NC_GLOBAL, name, NC_INT, 1, &ndims); ERR
+    for (i=0; i<ndims; i++) {
+        sprintf(name, "D%d.dim_%d", label, i);
+        /* dims[] is in Fortran order */
+        err = ncmpi_put_att_longlong(ncid, NC_GLOBAL, name, NC_INT, 1, &dims[ndims-1-i]);
         ERR
     }
-    else {
-        if ((ndims == 1 && dim_len != dims[0]) || (ndims == 2 && dim_len != dims[1])) {
-            printf("Error: dim_len_X mismatches among input files\n");
-            MPI_Finalize();
-            exit(1);
-        }
-    }
-    free(dims);
 
-    sprintf(name, "%s.max_nreqs", label);
+    sprintf(name, "D%d.max_nreqs", label);
     err = ncmpi_put_att_int(ncid, NC_GLOBAL, name, NC_INT, 1, &max_nreqs); ERR
-    sprintf(name, "%s.min_nreqs", label);
+    sprintf(name, "D%d.min_nreqs", label);
     err = ncmpi_put_att_int(ncid, NC_GLOBAL, name, NC_INT, 1, &min_nreqs); ERR
+
+    /* exit define mode */
     err = ncmpi_enddef(ncid); ERR
 
-    start[0] = 0;
-    count[0] = nprocs;
+    /* write variable containing number of requests for each process */
     err = ncmpi_put_var_int_all(ncid, varid[0], nreqs); ERR
 
-    off = (int*) malloc(max_nreqs * sizeof(int));
-
+    /* read the offsets again into allocated array off */
+    start = 0;
     rewind(fd);
     fgets(buf, LINE_SIZE, fd);
     fgets(buf, LINE_SIZE, fd);
     for (rank=0; rank<nprocs; rank++) {
-        fgets(buf, LINE_SIZE, fd);
-        assert(rank == atoi(strtok(buf, " ")));
-        nreqs[rank] = atoi(strtok(NULL, " "));
+        int prev, ncontig=0, decomp_rank;
+        char *ret;
+
+        /* reads the first line of rank ID and no. requests */
+        while (NULL != (ret = fgets(buf, LINE_SIZE, fd))) {
+            if (buf[0] != '\n') break; /* non-empty line */
+        }
+        if (ret == NULL || strncmp(buf, "Obtained", 8) == 0) {
+            /* there is no request for remaining ranks */
+            for (decomp_rank=rank; decomp_rank<nprocs; decomp_rank++)
+                nreqs[rank] = 0;
+            break; /* loop of rank */
+        }
+
+        decomp_rank = atoi(strtok(buf, " ")); /* rank ID */
+        while (rank < decomp_rank) /* this rank has no request */
+            rank++;
+
+        nreqs[rank] = atoi(strtok(NULL, " "));  /* number of requests */
+        if (nreqs[rank] == 0) { /* this rank has zero request */
+            min_nreqs = 0;
+            continue; /* loop of rank */
+        }
+
+        off = (int*) malloc(nreqs[rank] * sizeof(int));
+        len = (int*) malloc(nreqs[rank] * sizeof(int));
         fgets(buf, LINE_SIZE, fd);
         off[0] = atoi(strtok(buf, " "));
         j=1;
@@ -222,13 +318,35 @@ int add_decomp(int   ncid,
             k++;
         }
 
-        start[0] = rank;
-        start[1] = 0;
-        count[0] = 1;
-        count[1] = k;
-        err = ncmpi_put_vara_int_all(ncid, varid[1], start, count, off); ERR
+        /* sort off[] into an increasing order */
+        qsort((void*)off, k, sizeof(int), intcompare);
+
+        ncontig = 1;
+        prev = 0;
+        len[0] = 1;
+        for (j=1; j<k; j++) {
+            /* break contiguity at dimension boundaries or noncontiguous */
+            if (off[j] % dims[0] == 0 || off[j] > off[j-1] + 1) ncontig++;
+
+            if (off[j] % dims[0] == 0 || off[j] > off[prev] + len[prev]) {
+                prev++;
+                if (prev < j) off[prev] = off[j];
+                len[prev] = 1;
+            }
+            else len[prev]++;
+        }
+        assert(prev+1 == ncontig);
+
+        /* write/append to variables offsets and lengths */
+        count = ncontig;
+        err = ncmpi_put_vara_int_all(ncid, varid[1], &start, &count, off); ERR
+        err = ncmpi_put_vara_int_all(ncid, varid[2], &start, &count, len); ERR
+        start += ncontig;
+
+        free(off);
+        free(len);
     }
-    free(off);
+    free(dims);
 
 fn_exit:
     fclose(fd);
@@ -247,17 +365,19 @@ usage(char *argv0)
     "       -v               Verbose mode\n"
     "       -l num           max number of characters per line in input file\n"
     "       -o out_file      name of output netCDF file\n"
-    "       -1 input_file    name of 1st 1D decomposition file\n"
-    "       -2 input_file    name of 2nd 1D decomposition file\n"
-    "       -3 input_file    name of     2D decomposition file\n";
+    "       -1 input_file    name of 1st decomposition file\n"
+    "       -2 input_file    name of 2nd decomposition file\n"
+    "       -3 input_file    name of 3rd decomposition file\n"
+    "       -4 input_file    name of 4th decomposition file\n"
+    "       -5 input_file    name of 5th decomposition file\n"
+    "       -6 input_file    name of 6th decomposition file\n";
     fprintf(stderr, help, argv0);
 }
 
 /*----< main() >------------------------------------------------------------*/
 int main(int argc, char **argv) {
-    char *infname_D1=NULL, *infname_D2=NULL, *infname_D3=NULL, outfname[1024];
-    char cmd_line[4096];
-    int  i, rank, ncid, err, nerrs=0;
+    char *infname[6], outfname[1024], cmd_line[4096];
+    int  i, rank, ncid, num_decomp, dimid, err, nerrs=0;
 
     MPI_Init(&argc, &argv);
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
@@ -268,12 +388,13 @@ int main(int argc, char **argv) {
         strcat(cmd_line, " ");
     }
 
+    for (i=0; i<6; i++) infname[i] = NULL;
     outfname[0] = '\0';
     line_sz = LINE_SIZE;
     verbose = 0;
 
     /* get command-line arguments */
-    while ((i = getopt(argc, argv, "hvo:l:1:2:3:")) != EOF)
+    while ((i = getopt(argc, argv, "hvo:l:1:2:3:4:5:6:")) != EOF)
         switch(i) {
             case 'v': verbose = 1;
                       break;
@@ -281,11 +402,23 @@ int main(int argc, char **argv) {
                       break;
             case 'l': line_sz = atoi(optarg);
                       break;
-            case '1': infname_D1 = optarg;
+            case '1': infname[0] = optarg;
+                      num_decomp = 1;
                       break;
-            case '2': infname_D2 = optarg;
+            case '2': infname[1] = optarg;
+                      num_decomp = 2;
                       break;
-            case '3': infname_D3 = optarg;
+            case '3': infname[2] = optarg;
+                      num_decomp = 3;
+                      break;
+            case '4': infname[3] = optarg;
+                      num_decomp = 4;
+                      break;
+            case '5': infname[4] = optarg;
+                      num_decomp = 5;
+                      break;
+            case '6': infname[5] = optarg;
+                      num_decomp = 6;
                       break;
             case 'h':
             default:  if (rank==0) usage(argv[0]);
@@ -293,17 +426,16 @@ int main(int argc, char **argv) {
                       return 1;
         }
 
-    if (infname_D1 == NULL || infname_D2 == NULL || infname_D3 == NULL) {
-        /* 3 input data decomposition files are mandatory */
-        usage(argv[0]);
-        MPI_Finalize();
-        return 1;
+    if (verbose) {
+        for (i=0; i<6; i++) {
+            if (infname[i] == NULL) continue;
+            printf("input file %d = %s\n",i, infname[i]);
+        }
     }
-    if (verbose) printf("input files =%s %s %s\n",infname_D1,infname_D2,infname_D3);
 
     if (outfname[0] == '\0') {
-        strcpy(outfname, infname_D1);
-        char *ptr = strstr(infname_D1, ".dat");
+        strcpy(outfname, infname[0]);
+        char *ptr = strstr(infname[0], ".dat");
         if (ptr == NULL) {
             if (rank == 0) {
                 printf("Error at line %d: expected input file name extension is \".dat\"\n",__LINE__);
@@ -316,9 +448,14 @@ int main(int argc, char **argv) {
     }
     if (verbose) printf("output file name =%s\n",outfname);
 
-    if (strcmp(infname_D1, outfname) == 0) {
+    if ((infname[0] != NULL && strcmp(infname[0], outfname) == 0) ||
+        (infname[1] != NULL && strcmp(infname[1], outfname) == 0) ||
+        (infname[2] != NULL && strcmp(infname[2], outfname) == 0) ||
+        (infname[3] != NULL && strcmp(infname[3], outfname) == 0) ||
+        (infname[4] != NULL && strcmp(infname[4], outfname) == 0) ||
+        (infname[5] != NULL && strcmp(infname[5], outfname) == 0)) {
         if (rank == 0) {
-            printf("Error: input and output file names are the same\n");
+            printf("Error: input and output file names conflict\n");
             printf("Abort ...\n");
         }
         MPI_Finalize();
@@ -333,13 +470,18 @@ int main(int argc, char **argv) {
         nerrs++;
         goto fn_exit;
     }
+
+    /* add the number of decompositions */
+    err = ncmpi_def_dim(ncid, "num_decomp", num_decomp, &dimid); ERR
+
+    /* add command line used */
     err = ncmpi_put_att_text(ncid, NC_GLOBAL, "command_line", strlen(cmd_line), cmd_line); ERR
 
-    nerrs += add_decomp(ncid, infname_D3, "D3");
-    err = ncmpi_redef(ncid); ERR;
-    nerrs += add_decomp(ncid, infname_D2, "D2");
-    err = ncmpi_redef(ncid); ERR;
-    nerrs += add_decomp(ncid, infname_D1, "D1");
+    for (i=0; i<6; i++) {
+        if (infname[i] == NULL) continue;
+        nerrs += add_decomp(ncid, infname[i], i+1);
+        err = ncmpi_redef(ncid); ERR;
+    }
 
     err = ncmpi_close(ncid); ERR
