+++ datasets/f_case_h0.txt
@@ -1,4 +1,4 @@
-netcdf testfile_h0_vard {
+netcdf f_case_h0_vard {
 // file format: CDF-5 (big variables)
 dimensions:
 	ncol = 866 ;
