+++ datasets/f_case_h1.txt
@@ -1,4 +1,4 @@
-netcdf testfile_h1_vard {
+netcdf f_case_h1_vard {
 // file format: CDF-5 (big variables)
 dimensions:
 	ncol = 866 ;
